requirements.txt
transformers
torch
langchain-openai
load_dotenv
langchain
langchain-community
crewai
crewai-tools
langchain-google-genai
streamlit
langchain-groq
agentops
mcp
langgraph
langchain-mcp-adapters
chromadb
langchain-chroma
llama-index
llama-index-llms-google-genai
llama-index-tools-mcp
google-genai
wikipedia
sqlalchemy==2.0.39
psycopg2-binary==2.9.10
